+++ packages/angular-workspace/nimble-angular/src/directives/select/nimble-select.directive.ts
@@ -71,5 +71,13 @@ export class NimbleSelectDirective {
         this.renderer.setProperty(this.elementRef.nativeElement, 'loadingVisible', toBooleanProperty(value));
     }
 
+    public get requiredVisible(): boolean {
+        return this.elementRef.nativeElement.requiredVisible;
+    }
+
+    @Input('required-visible') public set requiredVisible(value: BooleanValueOrAttribute) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'requiredVisible', toBooleanProperty(value));
+    }
+
     public constructor(private readonly renderer: Renderer2, private readonly elementRef: ElementRef<Select>) {}
 }