+++ packages/angular-workspace/nimble-angular/src/directives/text-area/nimble-text-area.directive.ts
@@ -138,5 +138,13 @@ export class NimbleTextAreaDirective {
         this.renderer.setProperty(this.elementRef.nativeElement, 'spellcheck', toBooleanProperty(value));
     }
 
+    public get requiredVisible(): boolean {
+        return this.elementRef.nativeElement.requiredVisible;
+    }
+
+    @Input('required-visible') public set requiredVisible(value: BooleanValueOrAttribute) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'requiredVisible', toBooleanProperty(value));
+    }
+
     public constructor(private readonly renderer: Renderer2, private readonly elementRef: ElementRef<TextArea>) {}
 }